Edit Log Added
#comment: Can edit any logs. If tour is deleted, every related to it is deleted (also car/bike logs)

#duration: 8

+++ TourPlanner.Model/Log/CarData.cs
@@ -15,8 +15,9 @@ namespace TourPlanner.Model.Log
 
         public int FuelUsed { get; set; }
 
-        public int FuelCost { get; set; }
+        public decimal FuelCost { get; set; }
 
         public bool CaughtSpeeding { get; set; }
+
     }
 }
